+++ dune/istl/bcrsmatrix.hh
@@ -502,10 +502,6 @@ namespace Dune {
     //! The type for the statistics object returned by compress()
     typedef ::Dune::CompressionStatistics<size_type> CompressionStatistics;
 
-    //! increment block level counter
-    [[deprecated("Use free function blockLevel(). Will be removed after 2.8.")]]
-    static constexpr unsigned int blocklevel = blockLevel<B>()+1;
-
     //! we support two modes
     enum BuildMode {
       /**