+++ dune/istl/matrix.hh
@@ -588,10 +588,6 @@ namespace MatrixImp
     /** \brief Const iterator for the entries of each row */
     typedef typename row_type::const_iterator ConstColIterator;
 
-    //! The number of nesting levels the matrix contains.
-    [[deprecated("Use free function blockLevel(). Will be removed after 2.8.")]]
-    static constexpr auto blocklevel = blockLevel<T>()+1;
-
     /** \brief Create empty matrix */
     Matrix() : data_(0,0), cols_(0)
     {}